Refactored listener by moving ITestListener,IInvokedMethodListener & ISuiteListener into CustomListener
