+++ TestNG/practicePageTestNG.xml
@@ -2,18 +2,16 @@
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
 <suite name="Suite" parallel="tests" thread-count="2">
 <listeners>
-	<listener class-name="com.automation.customEventListener.IsuiteListener"></listener>
-	<listener class-name="com.automation.customEventListener.InvokedMethodListener"></listener>
-	<listener class-name="com.automation.customEventListener.ItestListener"></listener>
+	<listener class-name="com.automation.customEventListener.CustomListener"></listener>
 </listeners>
-  <test name="Test1">
+  <test name="ChromeTest">
 	  <parameter name="appURL" value="https://letskodeit.teachable.com/p/practice"></parameter>
 	  <parameter name="browserType" value="chrome"/>
     <classes>
       <class name="seleniumhybrid.PracticeTest"/>
     </classes>
   </test> <!-- Test -->
-   <test name="Test2">
+   <test name="FirefoxTest">
 	  <parameter name="appURL" value="https://letskodeit.teachable.com/p/practice"></parameter>
 	  <parameter name="browserType" value="firefox"/>
     <classes>
